+++ server/src/main/java/org/opensearch/indices/IndicesService.java
@@ -302,6 +302,18 @@ public class IndicesService extends AbstractLifecycleComponent
     );
 
     /**
+     * This setting is used to prevents creation of indices where the 'index.replication.type' setting does not match
+     * with the replication type index setting, set at cluster level i.e. 'cluster.indices.replication.strategy'.
+     * If disabled, the replication type can be specified.
+     */
+    public static final Setting<Boolean> CLUSTER_RESTRICT_INDEX_REPLICATION_TYPE_SETTING = Setting.boolSetting(
+        "cluster.force.index.replication_type",
+        false,
+        Property.NodeScope,
+        Property.Final
+    );
+
+    /**
      * The node's settings.
      */
     private final Settings settings;